Restructure basic recovery tests
