+++ dev/com.ibm.ws.transaction.recovery_fat.1/bnd.bnd
@@ -1,5 +1,5 @@
 #*******************************************************************************
-# Copyright (c) 2019, 2022 IBM Corporation and others.
+# Copyright (c) 2019, 2023 IBM Corporation and others.
 # All rights reserved. This program and the accompanying materials
 # are made available under the terms of the Eclipse Public License 2.0
 # which accompanies this distribution, and is available at
@@ -15,10 +15,10 @@ bVersion=1.0
 
 src: \
 	fat/src,\
-	test-applications/transaction/src,\
-	test-bundles/xaflowcallback/src
+	test-applications/transaction/src
 
--sub: *.bnd
+# Define the bundle for this FAT
+Bundle-SymbolicName: ${basename;${basedir}}
 
 fat.project: true
 
@@ -40,11 +40,6 @@ tested.features:\
   servlet-4.0, servlet-5.0, servlet-6.0,\
   xmlbinding-3.0, xmlbinding-4.0
 
-#  jpacontainer-2.2,\
-#  jpa-2.2,\
-#  txtest-2.0,\
-#  jdbc-4.3
-
 # To define a global minimum java level for the FAT, use the following property.
 # If unspecified, the default value is ${javac.source}
 # fat.minimum.java.level: 11
@@ -61,7 +56,6 @@ tested.features:\
   com.ibm.websphere.javaee.transaction.1.2;version=latest,\
   com.ibm.ws.logging,\
   com.ibm.ws.org.osgi.annotation.versioning;version=latest,\
-  com.ibm.ws.tx.embeddable;version=latest,\
+  com.ibm.ws.transaction.fat.util;version=latest,\
   com.ibm.ws.transaction.test.util;version=latest,\
-  com.ibm.ws.transaction.fat.util;version=latest
-  
+  com.ibm.ws.tx.embeddable;version=latest
