--- dev/com.ibm.ws.transaction.recovery_fat.1/xaflow.bnd
@@ -1,22 +1,0 @@
-#*******************************************************************************
-# Copyright (c) 2020 IBM Corporation and others.
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the Eclipse Public License 2.0
-# which accompanies this distribution, and is available at
-# http://www.eclipse.org/legal/epl-2.0/
-# 
-# SPDX-License-Identifier: EPL-2.0
-#
-# Contributors:
-#     IBM Corporation - initial API and implementation
-#*******************************************************************************
--include= ~../cnf/resources/bnd/bundle.props
-bVersion=1.0
-
-Bundle-Name: TransactionsXAFlowCallback
-Bundle-SymbolicName: com.ibm.ws.tx.test.impl
-Bundle-Description: XAFlowCallback classes for transactions FATs; version=${bVersion}
-
-WS-TraceGroup: Transaction
-
-Export-Package: com.ibm.ws.tx.test.impl;version=1.0
