--- dev/com.ibm.ws.transaction.recovery_fat.2/.gitignore
@@ -1,3 +1,0 @@
-/publish
-/test-applications
-/test-bundles
