+++ dev/com.ibm.ws.transaction.recovery_fat.2/bnd.bnd
@@ -1,5 +1,5 @@
 #*******************************************************************************
-# Copyright (c) 2019, 2022 IBM Corporation and others.
+# Copyright (c) 2019, 2023 IBM Corporation and others.
 # All rights reserved. This program and the accompanying materials
 # are made available under the terms of the Eclipse Public License 2.0
 # which accompanies this distribution, and is available at
@@ -13,12 +13,11 @@
 -include= ~../cnf/resources/bnd/bundle.props
 bVersion=1.0
 
-src: \
-	fat/src,\
-	test-applications/transaction/src,\
-	test-bundles/xaflowcallback/src
+# Define the bundle for this FAT
+Bundle-SymbolicName: ${basename;${basedir}}
 
--sub: *.bnd
+src: \
+	fat/src
 
 fat.project: true
 
@@ -52,7 +51,7 @@ tested.features:\
   com.ibm.websphere.javaee.transaction.1.2;version=latest,\
   com.ibm.ws.logging,\
   com.ibm.ws.org.osgi.annotation.versioning;version=latest,\
-  com.ibm.ws.tx.embeddable;version=latest,\
-  com.ibm.ws.transaction.test.util;version=latest,\
   com.ibm.ws.transaction.fat.util;version=latest,\
-  com.ibm.ws.transaction.recovery_fat.1;version=latest
+  com.ibm.ws.transaction.recovery_fat.1;version=latest,\
+  com.ibm.ws.transaction.test.util;version=latest,\
+  com.ibm.ws.tx.embeddable;version=latest
