+++ dev/com.ibm.ws.transaction.recovery_fat.2/build.gradle
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2021 IBM Corporation and others.
+ * Copyright (c) 2021, 2023 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
@@ -19,7 +19,6 @@ dependencies {
                project(':io.openliberty.org.apache.commons.logging')
 }
 
-
 File sharedDir = new File(autoFvtDir, 'publish/shared/resources')
 
 task addDerbyToSharedDir(type: Copy) {
@@ -28,61 +27,18 @@ task addDerbyToSharedDir(type: Copy) {
   rename 'derby-.*.jar', 'derby.jar'
 }
 
-task copyFAT {
-    copy {
-      from project(':com.ibm.ws.transaction.recovery_fat.1').file('fat/src/com/ibm/ws/transaction/test/tests')
-      into new File('fat/src/com/ibm/ws/transaction/test/tests')
-    }
+task addDerbyToSharedDir2(type: Copy) {
+  from configurations.derby
+  into new File(sharedDir, 'derby2')
+  rename 'derby-.*.jar', 'derby.jar'
 }
 
-task copyCommonFiles {
-  doLast {
-    /*
-     * Copy test application resources.
-     */
-    copy {
-      from project(':com.ibm.ws.transaction.recovery_fat.1').file('test-applications')
-      into new File('test-applications')
-    }
-    copy {
-      from project(':com.ibm.ws.transaction.recovery_fat.1').file('test-bundles')
-      into new File('test-bundles')
-    }
-
-    /*
-     * Copy common publish files.
-     */
-    copy {
-      from project(':com.ibm.ws.transaction.recovery_fat.1').file('publish')
-      into new File('publish')
-    }
-  }
-}
-
-addRequiredLibraries.dependsOn copyFAT
 addRequiredLibraries.dependsOn addDerbyToSharedDir
+addRequiredLibraries.dependsOn addDerbyToSharedDir2
 addRequiredLibraries.dependsOn addJakartaTransformer
 
-jar {
-  dependsOn copyCommonFiles
-}
-
-autoFVT.dependsOn copyCommonFiles
-
 clean.doLast {
-  if (file('test-applications').exists()){
-    file('test-applications').deleteDir()
-  }
-  if (file('test-bundles').exists()){
-    file('test-bundles').deleteDir()
-  }
-  if (file('publish').exists()){
-    file('publish').deleteDir()
-  }
   if (file('build').exists()){
     file('build').deleteDir()
-  }
-  if (file('fat/src/com/ibm/ws/transaction/test/tests').exists()){
-    file('fat/src/com/ibm/ws/transaction/test/tests').deleteDir()
   }
 }
