--- dev/com.ibm.ws.transaction.recovery_fat.2/com.ibm.ws.transactions.fat.utils.bnd
@@ -1,26 +1,0 @@
-#*******************************************************************************
-# Copyright (c) 2018 IBM Corporation and others.
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the Eclipse Public License 2.0
-# which accompanies this distribution, and is available at
-# http://www.eclipse.org/legal/epl-2.0/
-# 
-# SPDX-License-Identifier: EPL-2.0
-#
-# Contributors:
-#     IBM Corporation - initial API and implementation
-#*******************************************************************************
--include= ~../cnf/resources/bnd/bundle.props
-bVersion=1.0
-
-Bundle-Name: TransactionsFATUtils
-Bundle-SymbolicName: com.ibm.ws.transactions.fat.utils
-Bundle-Description: Utility classes for transactions FATs; version=${bVersion}
-Web-ContextPath: inflow
-
-WS-TraceGroup: Transaction
-
-Export-Package: com.ibm.tx.jta.ut.util
-
--dsannotations: \
- com.ibm.tx.jta.ut.util.XAResourceFactoryService
