+++ dev/com.ibm.ws.transaction.recovery_fat.2/fat/src/tests/RecoveryDBTranlogTest.java
@@ -1,0 +1,38 @@
+/*******************************************************************************
+ * Copyright (c) 2017, 2023 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License 2.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     IBM Corporation - initial API and implementation
+ *******************************************************************************/
+package tests;
+
+import org.junit.BeforeClass;
+import org.junit.runner.RunWith;
+
+import com.ibm.websphere.simplicity.log.Log;
+
+import componenttest.annotation.Server;
+import componenttest.custom.junit.runner.FATRunner;
+import componenttest.custom.junit.runner.Mode;
+import componenttest.custom.junit.runner.Mode.TestMode;
+import componenttest.topology.impl.LibertyServer;
+
+@Mode(TestMode.FULL)
+@RunWith(FATRunner.class)
+public class RecoveryDBTranlogTest extends RecoveryTestBase {
+
+    @Server("recovery.dblog")
+    public static LibertyServer s;
+
+    @BeforeClass
+    public static void setup() throws Exception {
+        Log.info(RecoveryTest.class, "subBefore", s.getServerName());
+        setup(s);
+    }
+}
