+++ dev/com.ibm.ws.transaction.recovery_fat.2/publish/servers/recovery.dblog/bootstrap.properties
