+++ dev/com.ibm.ws.transaction.recovery_fat.3/.classpath
@@ -1,9 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <classpath>
-	<classpathentry kind="src" path="fat/src" including="com/ibm/ws/transaction/test/FATSuite.java"/>
-	<classpathentry kind="src" path="com.ibm.ws.transaction.recovery_fat.1-fatsrc" excluding="com/ibm/ws/transaction/test/FATSuite.java"/>
-	<classpathentry kind="src" path="com.ibm.ws.transaction.recovery_fat.1-xaflowcallback"/>
-	<classpathentry kind="src" path="com.ibm.ws.transaction.recovery_fat.1-transaction"/>
+	<classpathentry kind="src" path="test-applications/transaction/src"/>
+	<classpathentry kind="src" path="fat/src"/>
+	<classpathentry kind="src" path="test-bundles/xaflowcallback/src"/>
 	<classpathentry exported="true" kind="con" path="aQute.bnd.classpath.container"/>
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8">
 		<attributes>
