+++ dev/com.ibm.ws.transaction.recovery_fat.3/bnd.bnd
@@ -18,6 +18,9 @@ src: \
 	test-applications/transaction/src,\
 	test-bundles/xaflowcallback/src
 
+# Define the bundle for this FAT
+Bundle-SymbolicName: ${basename;${basedir}}
+
 -sub: *.bnd
 
 fat.project: true
@@ -56,7 +59,7 @@ tested.features:\
   com.ibm.websphere.javaee.transaction.1.2;version=latest,\
   com.ibm.ws.logging,\
   com.ibm.ws.org.osgi.annotation.versioning;version=latest,\
-  com.ibm.ws.tx.embeddable;version=latest,\
-  com.ibm.ws.transaction.test.util;version=latest,\
   com.ibm.ws.transaction.fat.util;version=latest,\
-  com.ibm.ws.transaction.recovery_fat.1;version=latest
+  com.ibm.ws.transaction.recovery_fat.1;version=latest,\
+  com.ibm.ws.transaction.test.util;version=latest,\
+  com.ibm.ws.tx.embeddable;version=latest
