--- dev/com.ibm.ws.transaction.recovery_fat.3/fat/src/com/ibm/ws/transaction/test/.gitignore
@@ -1,1 +1,0 @@
-/tests
