--- dev/com.ibm.ws.transaction.recovery_fat.3/fat/src/com/ibm/ws/transaction/test/FATSuite.java
@@ -1,40 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2017, 2022 IBM Corporation and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License 2.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-2.0/
- * 
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *     IBM Corporation - initial API and implementation
- *******************************************************************************/
-package com.ibm.ws.transaction.test;
-
-import org.junit.ClassRule;
-import org.junit.runner.RunWith;
-import org.junit.runners.Suite;
-import org.junit.runners.Suite.SuiteClasses;
-
-import com.ibm.ws.transaction.test.tests.DataSourceChangeTest;
-import com.ibm.ws.transaction.test.tests.WaitForRecoveryTest;
-import com.ibm.ws.transaction.test.tests.XAFlowTest;
-
-import componenttest.rules.repeater.FeatureReplacementAction;
-import componenttest.rules.repeater.RepeatTests;
-
-@RunWith(Suite.class)
-@SuiteClasses({
-                XAFlowTest.class,
-                WaitForRecoveryTest.class,
-                DataSourceChangeTest.class,
-})
-public class FATSuite {
-
-    @ClassRule
-    public static RepeatTests r = RepeatTests.withoutModification()
-                    .andWith(FeatureReplacementAction.EE8_FEATURES().fullFATOnly())
-                    .andWith(FeatureReplacementAction.EE9_FEATURES().fullFATOnly())
-                    .andWith(FeatureReplacementAction.EE10_FEATURES().fullFATOnly());
-}
