+++ dev/com.ibm.ws.transaction.recovery_fat.3/publish/servers/com.ibm.ws.transaction_XAFlow/resources/security/defaultTrust.jks
